+++ rocketgram/__init__.py
@@ -8,7 +8,8 @@ from . import tools
 from .api import *
 from .bot import Bot
 from .connectors import *
-from .context import context2
+from .context import context
+from .context import context as context2
 from .errors import *
 from .executors import *
 from .keyboards import *