+++ rocketgram/api/add_sticker_to_set.py
@@ -6,9 +6,8 @@
 from dataclasses import dataclass
 from typing import Union, Optional, List
 
-from .input_file import InputFile
-from .mask_position import MaskPosition
 from .request import Request
+from .. import api
 
 
 @dataclass(frozen=True)
@@ -22,14 +21,14 @@ class AddStickerToSet(Request):
 
     user_id: int
     name: str
-    png_sticker: Optional[Union[InputFile, str]]
-    tgs_sticker: Optional[InputFile]
+    png_sticker: Optional[Union['api.InputFile', str]]
+    tgs_sticker: Optional['api.InputFile']
     emojis: str
-    mask_position: Optional[MaskPosition] = None
+    mask_position: Optional['api.MaskPosition'] = None
 
-    def files(self) -> List[InputFile]:
+    def files(self) -> List['api.InputFile']:
         files = list()
-        if isinstance(self.png_sticker, InputFile):
+        if isinstance(self.png_sticker, api.InputFile):
             files.append(self.png_sticker)
         if self.tgs_sticker is not None:
             files.append(self.tgs_sticker)